+++ packages/terra-data-grid/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 ## Unreleased
+* Fixed
+  * Fixed issue where WorklistDataGrid loses focus when Row Selection Mode is turned off from a Row Selection Cell.
+  * Fixed tab focus management when next element is in a hidden container.
+  * Fixed tab focus management to wrap to the first focusable element when the data grid is the last focusable element.
+
+## 0.8.0 - (October 11, 2023)
 
 * Added
   * Added base FlowsheetDataGrid component.